+++ src/constants/NavigationMeta.js
@@ -182,7 +182,7 @@ export const PRIVATE_NAV = [
   },
   {
     icon: 'ns-article',
-    title: { issuer: 'Resources', investor: 'Education Centre' },
+    title: { issuer: 'Resources', investor: 'Education Center' },
     to: 'resources',
     accessibleTo: ['investor', 'issuer'],
     subPanel: 1,
@@ -280,8 +280,20 @@ export const PRIVATE_NAV = [
     icon: 'dollar',
     title: 'CrowdPay',
     to: 'crowdPay',
+    heading: 'Manage Crowdpay Accounts',
     path: 'admin/crowdPay',
     accessibleTo: ['admin'],
+    subPanel: 0,
+    subNavigations: [
+      { title: 'Review', to: 'review', component: 'ApplicationsList' },
+      { title: 'CIP', to: 'cip', component: 'ApplicationsList' },
+      { title: 'IRA', to: 'ira', component: 'ApplicationsList' },
+      {
+        title: 'Entity',
+        to: 'entity',
+        component: 'ApplicationsList',
+      },
+    ],
   },
   {
     icon: 'credit card',