+++ src/modules/private/investor/accountSetup/containers/accountCreation/entity/General.js
@@ -1,12 +1,16 @@
 import React, { Component } from 'react';
-import { Header, Form } from 'semantic-ui-react';
+import { Header, Form, Message } from 'semantic-ui-react';
 import { inject, observer } from 'mobx-react';
 import { US_STATES } from '../../../../../../../constants/account';
+import { ListErrors } from '../../../../../../../theme/shared';
 import { FormInput, MaskedInput, FormSelect, AutoComplete } from '../../../../../../../theme/form';
 
-@inject('entityAccountStore')
+@inject('entityAccountStore', 'uiStore')
 @observer
 export default class General extends Component {
+  componentWillMount() {
+    this.props.uiStore.setErrors(null);
+  }
   render() {
     const {
       GEN_INFO_FRM,
@@ -14,9 +18,15 @@ export default class General extends Component {
       setAddressFields,
       maskedGenInfoChange,
     } = this.props.entityAccountStore;
+    const { errors } = this.props.uiStore;
     return (
       <div>
         <Header as="h3" textAlign="center">General information</Header>
+        {errors &&
+          <Message error textAlign="left">
+            <ListErrors errors={[errors]} />
+          </Message>
+        }
         <Form error>
           <div className="field-wrap">
             <FormInput