+++ src/modules/private/shared/settings/components/ProfileData.js
@@ -12,7 +12,7 @@ import NewPhoneNumber from './profileSettings/NewPhoneNumber';
 import NewEmailAddress from './profileSettings/NewEmailAddress';
 import UpdateProfilePhoto from './profileSettings/UpdateProfilePhoto';
 import Helper from '../../../../../helper/utility';
-import { Spinner, UserAvatar } from '../../../../../theme/shared';
+import { InlineLoader, UserAvatar } from '../../../../../theme/shared';
 
 @inject('userDetailsStore', 'userStore', 'identityStore', 'uiStore')
 @observer
@@ -40,6 +40,7 @@ export default class ProfileData extends Component {
       email, legalDetails, avatar, firstName, lastName,
     } = this.props.userDetailsStore.userDetails;
     const User = { ...this.props.userStore.currentUser };
+    console.log(this.props.userDetailsStore.userDetails, User);
     const userAvatar = {
       firstName, lastName, avatarUrl: avatar ? avatar.url : '', roles: toJS(User.roles),
     };
@@ -49,11 +50,7 @@ export default class ProfileData extends Component {
       setAddressFieldsForProfile,
     } = this.props.identityStore;
     if (isEmpty(this.props.userDetailsStore.userDetails)) {
-      return (
-        <div>
-          <Spinner loaderMessage="Loading..." />
-        </div>
-      );
+      return <InlineLoader />;
     }
     return (
       <Grid>
@@ -137,12 +134,14 @@ export default class ProfileData extends Component {
                 <Link to={`${this.props.match.url}/update-profile-photo`}><b>Change profile photo</b></Link>
               </div>
             </Card>
-            <UserVerifiedDetails
-              {...this.props}
-              email={email}
-              legalDetails={legalDetails}
-              isUserVerified={this.isVerified}
-            />
+            {userAvatar.roles.includes('investor') &&
+              <UserVerifiedDetails
+                {...this.props}
+                email={email}
+                legalDetails={legalDetails}
+                isUserVerified={this.isVerified}
+              />
+            }
           </Card.Group>
         </Grid.Column>
       </Grid>