+++ src/services/stores/queries/account.js
@@ -73,3 +73,10 @@ export const updateInvestorProfileData = gql`
       }
     }
   }`;
+
+export const checkEntityTaxIdCollision = gql`
+query checkEntityTaxIdCollision($taxId: String!) {
+  checkEntityTaxIdCollision(taxId: $taxId) {
+    alreadyExists
+  }
+}`;